+++ clase16/ejercicio clase 16/ejercicioClase16/ejercicioClase16/Modelos/Cuadrilatero.cs
@@ -62,6 +62,6 @@ namespace ejercicioClase16.Modelos
             Vertice3 = v3;
             Vertice4 = v4;
         }
-        public abstract double CalcularArea();
+        public abstract decimal CalcularArea();
     }
 }
